Update model classes

+++ LibrarySystem-Backend/src/main/java/ca/mcgill/ecse321/library/dao/ReservationRepository.java
@@ -8,9 +8,9 @@ import ca.mcgill.ecse321.library.models.LibraryItem;
 import ca.mcgill.ecse321.library.models.Reservation;
 import ca.mcgill.ecse321.library.models.User;
 
-public interface ReservationRepository extends CrudRepository<Reservation, Integer>{
+public interface ReservationRepository extends CrudRepository<Reservation, Long>{
 	
-	Reservation findReservationByReservationID(int reservationID);
+	Reservation findReservationByReservationID(Long reservationID);
 	Reservation findByLibraryItem(LibraryItem libraryItem);
 	List<Reservation> findByUser(User user);
